Added base for tutorial... Added some functional for techtree... Added help.html for login window

+++ source/monitor/NrpDevelopGame.cpp
@@ -86,14 +86,14 @@ CNrpDevelopGame::CNrpDevelopGame( CNrpGameProject* nProject, CNrpCompany* ptrCom
 	if( nProject->GetValue<PNrpTechnology>( SCRIPTENGINE ) )
 	{
 		PNrpProjectModule module = new CNrpProjectModule( nProject->GetValue<PNrpTechnology>( SCRIPTENGINE ), this );
-		module->SetValue<int>( CODEVOLUME, bcv * module->GetValue<float>( BASE_CODE ) );
+		module->SetValue<int>( CODEVOLUME, static_cast< int >( bcv * module->GetValue<float>( BASE_CODE ) ) );
 		gameModules_.push_back( module );
 	}
 
 	if( nProject->GetValue<PNrpTechnology>( MINIGAMEENGINE ) )
 	{
 		PNrpProjectModule module = new CNrpProjectModule( nProject->GetValue<PNrpTechnology>( MINIGAMEENGINE ), this );
-		module->SetValue<int>( CODEVOLUME, bcv * module->GetValue<float>( BASE_CODE ) );
+		module->SetValue<int>( CODEVOLUME, static_cast< int >( bcv * module->GetValue<float>( BASE_CODE ) ) );
 		gameModules_.push_back( module );
 	}
